chore: update dependencies and improve Dart SDK compatibility
- Bump versions of several dependencies in package_graph.json including:
  - modular_core to 3.4.1
  - flutter_modular to 6.4.1
  - test to 1.30.0
  - meta to 1.17.0
  - coverage to 1.15.1
  - analyzer to 10.0.1
  - leak_tracker to 11.0.2
  - matcher to 0.12.19
  - material_color_utilities to 0.13.0
  - yaml to 3.1.3
- Add new web dependency in flutter_modular.
- Remove deprecated dependencies and files.
- Update GitHub Actions workflows to use latest actions and Java version.
- Modify pubspec.yaml files to require Dart SDK >=3.11.0.
- Introduce new web_url_service.dart for web URL handling.
- Update modular_base.dart and modular_route_information_parser.dart for improved route handling.
- Adjust tests to reflect changes in route handling logic.
