+++ .dart_tool/extension_discovery/vs_code.json
@@ -1,0 +1,1 @@
+{"version":2,"entries":[{"package":"flutter_modular_workspace","rootUri":"../","packageUri":"lib/"},{"package":"modular_core","rootUri":"../modular_core/","packageUri":"lib/"},{"package":"flutter_modular","rootUri":"../flutter_modular/","packageUri":"lib/"},{"package":"shelf_modular","rootUri":"../shelf_modular/","packageUri":"lib/"},{"package":"flutter_modular_example","rootUri":"../flutter_modular/example/","packageUri":"lib/"},{"package":"shelf_modular_example","rootUri":"../shelf_modular/example/","packageUri":"lib/"}]}
