+++ .dart_tool/pub/workspace_ref.json
@@ -1,3 +1,3 @@
 {
-  "workspaceRoot": "..\\.."
+  "workspaceRoot": "../.."
 }
