--- flutter_modular/lib/src/infra/services/url_service/html_url_service.dart
@@ -1,30 +1,0 @@
-// ignore_for_file: avoid_web_libraries_in_flutter
-
-import 'dart:html';
-
-import 'package:flutter_modular/flutter_modular.dart';
-import 'package:flutter_web_plugins/url_strategy.dart';
-
-import 'url_service.dart';
-
-class WebUrlService extends UrlService {
-  @override
-  String? getPath() {
-    final href = window.location.href;
-
-  if (urlStrategy is HashUrlStrategy) {
-      if (href.contains('#')) {
-        return href.split('#').last;
-      } else if (href.endsWith(Modular.initialRoute)) {
-        return Modular.initialRoute;
-        
-      }
-    }
-
-    return resolvePath(href);
-  }
-}
-
-UrlService create() {
-  return WebUrlService();
-}
