+++ flutter_modular/pubspec.yaml
@@ -6,12 +6,13 @@ homepage: https://github.com/Flutterando/modular
 resolution: workspace
 
 environment:
-  sdk: ">=3.6.0 <4.0.0"
+  sdk: ">=3.11.0 <4.0.0"
 
 dependencies:
   modular_core: ">=3.4.1 <4.0.0"
   meta: ">=1.3.0 <2.0.0"
   result_dart: ">=2.1.0 <3.0.0"
+  web: ">=0.5.0 <2.0.0"
   flutter:
     sdk: flutter
   flutter_web_plugins:
