+++ pubspec.yaml
@@ -1,7 +1,7 @@
 name: flutter_modular_workspace
 publish_to: none
 environment:
-  sdk: ">=3.6.0 <4.0.0"
+  sdk: ">=3.11.0 <4.0.0"
 
 workspace:
   - modular_core
